Fix climate icon color
