+++ src/common/entity/color/climate_color.ts
@@ -1,3 +1,13 @@
+import { HvacAction } from "../../../data/climate";
+
+export const CLIMATE_HVAC_ACTION_COLORS: Record<HvacAction, string> = {
+  cooling: "var(--rgb-state-climate-cool-color)",
+  drying: "var(--rgb-state-climate-dry-color)",
+  heating: "var(--rgb-state-climate-heat-color)",
+  idle: "var(--rgb-state-climate-idle-color)",
+  off: "var(--rgb-state-climate-off-color)",
+};
+
 export const climateColor = (state: string): string | undefined => {
   switch (state) {
     case "auto":
