+++ src/panels/lovelace/cards/tile/badges/tile-badge-climate.ts
@@ -5,16 +5,9 @@ import {
   mdiSnowflake,
   mdiWaterPercent,
 } from "@mdi/js";
+import { CLIMATE_HVAC_ACTION_COLORS } from "../../../../../common/entity/color/climate_color";
 import { ClimateEntity, HvacAction } from "../../../../../data/climate";
 import { ComputeBadgeFunction } from "./tile-badge";
-
-export const CLIMATE_HVAC_ACTION_COLORS: Record<HvacAction, string> = {
-  cooling: "var(--rgb-state-climate-cool-color)",
-  drying: "var(--rgb-state-climate-dry-color)",
-  heating: "var(--rgb-state-climate-heat-color)",
-  idle: "var(--rgb-state-climate-idle-color)",
-  off: "var(--rgb-state-climate-off-color)",
-};
 
 export const CLIMATE_HVAC_ACTION_ICONS: Record<HvacAction, string> = {
   cooling: mdiSnowflake,
